Move excel writer to new io package
