+++ pdtable/__init__.py
@@ -10,9 +10,8 @@ from .store import TableBundle, BlockType, BlockGenerator
 from .units import UnitPolicy
 from .io.parsers.fixer import ParseFixer
 from .io import read_csv
-from .io import read_excel
+from .io import read_excel, write_excel
 from .io.parsers.blocks import JsonData
 from .writers import write_csv
-from .writers import write_excel
 from .utils import read_bundle_from_csv, normalized_table_generator
 from .json_utils import json_data_to_table, table_to_json_data
