+++ pdtable/io/__init__.py
@@ -1,2 +1,2 @@
 from ._csv import read_csv
-from ._excel import read_excel
+from ._excel import read_excel, write_excel
