+++ pdtable/writers/__init__.py
@@ -1,2 +1,1 @@
 from ._csv import write_csv
-from ._excel import write_excel
